Created build command
